Change the way that dataset readme is handled
Dataset readme files tend to be in YAML format. However, this
is a convention rather than a prescription.

When generating the information to upload a dataset to the
lookup server the `utils.generate_dataset_info` converted the
YAML text to a dictionary and it was registered as such in the
mongo database. This would result in errors if the readme contained
invalid YAML.

It would be better if the base system deal with the readme as a
string of text rather than trying to interpret it.